Models juan

+++ src/main/java/com/workable_sb/workable/repositories/TipoContratoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.workable_sb.workable.models.TipoContrato;
 
-public interface TipoContratoRepository extends JpaRepository<TipoContrato, Short> {
+public interface TipoContratoRepository extends JpaRepository<TipoContrato, Integer> {
 
 }
